Rename `src/components/bottom-tabs` directory
