+++ src/components/tabs/TabsAccessory.types.ts
