+++ src/index.tsx
@@ -62,4 +62,4 @@ export { default as useTransitionProgress } from './useTransitionProgress';
 /**
  * EXPERIMENTAL API BELOW. MIGHT CHANGE W/O ANY NOTICE
  */
-export { default as Tabs } from './components/bottom-tabs';
+export { default as Tabs } from './components/tabs';
